# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Big Exec Pro Football Machine API",
    "version": "2026-09-08",
    "description": "Authenticated read-only machine surfaces over existing Big Exec beta operations."
  },
  "servers": [
    {
      "url": "https://www.bigexecfs.com"
    }
  ],
  "paths": {
...
    "/api/assistant-gm/tools": {
      "post": {
        "operationId": "runAssistantGmReadTools",
        "summary": "Run declared read-only Front Office Advisor tools for the signed-in user.",
        "responses": {
          "200": {
            "description": "Tool responses."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Policy denial."
          }
        }
      }
    },
    "/api/assistant-gm/run": {
      "post": {
        "operationId": "runAssistantGmAgentLoop",
        "summary": "Run a bounded, policy-gated Front Office Advisor read-tool loop for the signed-in user.",
        "responses": {
          "200": {
            "description": "Completed bounded run."
          },
          "400": {
            "description": "Invalid request."
          },
          "401": {
            "description": "Not signed in."
          },
          "403": {
            "description": "Policy denial."
          }
        }
      }
    },
...
  }
}